apps/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "primust-api"
version = "0.0.1"
description = "Primust control plane API (api.primust.com)"
requires-python = ">=3.11"
license = "LicenseRef-Proprietary"
dependencies = [
  "fastapi[standard]>=0.115.0",
  "uvicorn>=0.30.0",
  "asyncpg>=0.30.0",
  "pyjwt[crypto]>=2.9.0",
  "httpx>=0.27.0",
  "structlog>=24.0.0",
  "google-cloud-kms>=2.21.0",
  "reportlab>=4.1.0",
  "pypdf>=4.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.24.0",
  "aiosqlite>=0.20.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/primust_api"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
